Handle non-existing redirects.yml

+++ kit/postbuild.sh
@@ -3,7 +3,11 @@ set -euo pipefail
 IFS=$'\n\t'
 
 cp src/routes/_toctree.yml build/_toctree.yml
-cp src/routes/_redirects.yml build/_redirects.yml
+
+# Copy redirects yml file if exists
+if [[ -f "src/routes/_redirects.yml" ]]; then
+    cp src/routes/_redirects.yml build/_redirects.yml
+fi
 
 # To avoid conflict with Hub Tailwind CSS build, 
 # 1. making doc-builder's PostCSS geenrated file an empty one
